+++ packages/bijou/src/index.ts
@@ -125,8 +125,11 @@ export {
   type StepperOptions,
   dag,
   dagSlice,
+  dagLayout,
   type DagNode,
   type DagOptions,
+  type DagNodePosition,
+  type DagLayout,
 } from './core/components/index.js';
 
 // Forms